# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/dividends:
    get:
      tags:
        - Dividends
      summary: List Dividends
      description: Get a paginated list of Dividends with optional filtering and sorting
      operationId: listDividends
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DividendList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Dividends
      summary: Create Dividend
      description: Create a new Dividend
      operationId: createDividend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDividendRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dividend'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/share_classes:
    get:
      tags:
        - ShareClasses
      summary: List ShareClasses
      description: Get a paginated list of ShareClasses with optional filtering and sorting
      operationId: listShareClasses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareClassList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ShareClasses
      summary: Create ShareClass
      description: Create a new ShareClass
      operationId: createShareClass
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareClassRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareClass'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Dividend:
      type: object
      required:
        - id
        - company_id
        - share_class_id
        - declaration_date
        - per_share_amount
        - shares_outstanding
        - total_amount
        - status
        - retained_earnings_account_id
        - dividend_payable_account_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        share_class_id:
          type: string
          format: uuid
        declaration_date:
          type: string
          format: date
        payment_date:
          nullable: true
          type: string
          format: date
        per_share_amount:
          type: string
        shares_outstanding:
          type: string
        total_amount:
          type: string
        status:
          type: string
        retained_earnings_account_id:
          type: string
          format: uuid
        dividend_payable_account_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateDividendRequest:
      type: object
      required:
        - company_id
        - share_class_id
        - declaration_date
        - per_share_amount
        - shares_outstanding
        - total_amount
        - status
        - retained_earnings_account_id
        - dividend_payable_account_id
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        share_class_id:
          type: string
          format: uuid
        declaration_date:
          type: string
          format: date
        payment_date:
          nullable: true
          type: string
          format: date
        per_share_amount:
          type: string
        shares_outstanding:
          type: string
        total_amount:
          type: string
        status:
          type: string
        retained_earnings_account_id:
          type: string
          format: uuid
        dividend_payable_account_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    DividendList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Dividend'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ShareClass:
      type: object
      required:
        - id
        - company_id
        - code
        - name
        - par_value
        - currency
        - share_capital_account_id
        - share_premium_account_id
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 120
        par_value:
          type: string
        currency:
          type: string
          maxLength: 3
        share_capital_account_id:
          type: string
          format: uuid
        share_premium_account_id:
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    CreateShareClassRequest:
      type: object
      required:
        - company_id
        - code
        - name
        - par_value
        - currency
        - share_capital_account_id
        - share_premium_account_id
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 120
        par_value:
          type: string
        currency:
          type: string
          maxLength: 3
        share_capital_account_id:
          type: string
          format: uuid
        share_premium_account_id:
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    ShareClassList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ShareClass'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []