# src/exports/types.rs
//! Public types for Equity module
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! These DTOs are the ONLY types other modules should use.
//! They are decoupled from internal domain entities.

use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc, NaiveDate};
use rust_decimal::Decimal;
use crate::domain::entity::*;

// ============================================================================
// DIVIDEND TYPES
// ============================================================================

/// Type-safe ID for Dividend
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct DividendId(pub Uuid);

impl DividendId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for DividendId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<DividendId> for Uuid {
    fn from(id: DividendId) -> Self {
        id.0
    }
}

/// Data transfer object for Dividend
///
/// This is the public representation of Dividend for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct DividendDto {
    pub id: DividendId,
    pub company_id: Uuid,
    pub share_class_id: Uuid,
    pub declaration_date: NaiveDate,
    pub payment_date: Option<NaiveDate>,
    pub per_share_amount: Decimal,
    pub shares_outstanding: Decimal,
    pub total_amount: Decimal,
    pub status: DividendStatus,
    pub retained_earnings_account_id: Uuid,
    pub dividend_payable_account_id: Uuid,
    pub metadata: serde_json::Value,
}

/// Summary view of Dividend for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct DividendSummary {
    pub id: DividendId,
    pub status: DividendStatus,
}

/// Reference to Dividend for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct DividendRef {
    pub id: DividendId,
}

// ============================================================================
// SHARECLASS TYPES
// ============================================================================

/// Type-safe ID for ShareClass
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct ShareClassId(pub Uuid);

impl ShareClassId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for ShareClassId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ShareClassId> for Uuid {
    fn from(id: ShareClassId) -> Self {
        id.0
    }
}

/// Data transfer object for ShareClass
///
/// This is the public representation of ShareClass for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ShareClassDto {
    pub id: ShareClassId,
    pub company_id: Uuid,
    pub code: String,
    pub name: String,
    pub par_value: Decimal,
    pub currency: String,
    pub share_capital_account_id: Uuid,
    pub share_premium_account_id: Uuid,
    pub status: ShareClassStatus,
    pub metadata: serde_json::Value,
}

/// Summary view of ShareClass for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ShareClassSummary {
    pub id: ShareClassId,
    pub name: String,
    pub status: ShareClassStatus,
}

/// Reference to ShareClass for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ShareClassRef {
    pub id: ShareClassId,
}

// ============================================================================
// SHAREHOLDER TYPES
// ============================================================================

/// Type-safe ID for Shareholder
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct ShareholderId(pub Uuid);

impl ShareholderId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for ShareholderId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ShareholderId> for Uuid {
    fn from(id: ShareholderId) -> Self {
        id.0
    }
}

/// Data transfer object for Shareholder
///
/// This is the public representation of Shareholder for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ShareholderDto {
    pub id: ShareholderId,
    pub company_id: Uuid,
    pub party_id: Option<Uuid>,
    pub name: String,
    pub holder_type: HolderType,
    pub metadata: serde_json::Value,
}

/// Summary view of Shareholder for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ShareholderSummary {
    pub id: ShareholderId,
    pub name: String,
}

/// Reference to Shareholder for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ShareholderRef {
    pub id: ShareholderId,
}

// ============================================================================
// SHARETRANSACTION TYPES
// ============================================================================

/// Type-safe ID for ShareTransaction
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct ShareTransactionId(pub Uuid);

impl ShareTransactionId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for ShareTransactionId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ShareTransactionId> for Uuid {
    fn from(id: ShareTransactionId) -> Self {
        id.0
    }
}

/// Data transfer object for ShareTransaction
///
/// This is the public representation of ShareTransaction for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ShareTransactionDto {
    pub id: ShareTransactionId,
    pub company_id: Uuid,
    pub share_class_id: Uuid,
    pub shareholder_id: Uuid,
    pub txn_type: ShareTxnType,
    pub quantity: Decimal,
    pub price_per_share: Decimal,
    pub amount: Decimal,
    pub counterparty_shareholder_id: Option<Uuid>,
    pub transfer_group_id: Option<Uuid>,
    pub posting_reference: Option<String>,
    pub txn_date: NaiveDate,
    pub gl_posted: bool,
    pub metadata: serde_json::Value,
}

/// Summary view of ShareTransaction for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ShareTransactionSummary {
    pub id: ShareTransactionId,
}

/// Reference to ShareTransaction for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ShareTransactionRef {
    pub id: ShareTransactionId,
}

// ============================================================================
// CUSTOM TYPES
// ============================================================================

// <<< CUSTOM TYPES START >>>
// Add custom public types here
// <<< CUSTOM TYPES END >>>