# src/presentation/http/dividend_handler.rs
//! Dividend REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{NaiveDate};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{DividendService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateDividendDto, UpdateDividendDto, PatchDividendDto, DividendResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum DividendError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for DividendError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for DividendError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "DIVIDEND_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "DIVIDEND_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "DIVIDEND_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "DIVIDEND_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for Dividend.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /dividends | List with pagination |
/// | POST | /dividends | Create new |
/// | GET | /dividends/:id | Get by ID |
/// | PUT | /dividends/:id | Full update |
/// | PATCH | /dividends/:id | Partial update |
/// | DELETE | /dividends/:id | Soft delete |
/// | POST | /dividends/bulk | Bulk create |
/// | POST | /dividends/upsert | Upsert |
/// | GET | /dividends/trash | List deleted |
/// | POST | /dividends/:id/restore | Restore |
/// | DELETE | /dividends/empty | Empty trash |
/// | GET | /dividends/:id/deleted | Get deleted by ID |
/// | DELETE | /dividends/trash/:id | Permanent delete from trash |
/// | GET | /dividends/count | Count active entities |
/// | GET | /dividends/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(DividendService::with_repository(repository));
/// let router = create_dividend_routes(service);
/// ```
pub fn create_dividend_routes(service: Arc<DividendService>) -> Router {
    BackboneCrudHandler::<DividendService, Dividend, CreateDividendDto, UpdateDividendDto, DividendResponseDto>::routes(
        service,
        "/dividends",
    )
}

/// Create Axum router with only the read (GET) endpoints for Dividend.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_dividend_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_dividend_read_routes(service: Arc<DividendService>) -> Router {
    BackboneCrudHandler::<DividendService, Dividend, CreateDividendDto, UpdateDividendDto, DividendResponseDto>::read_routes(
        service,
        "/dividends",
    )
}

/// Create Axum router with only the write (mutation) endpoints for Dividend.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_dividend_write_routes(service: Arc<DividendService>) -> Router {
    BackboneCrudHandler::<DividendService, Dividend, CreateDividendDto, UpdateDividendDto, DividendResponseDto>::write_routes(
        service,
        "/dividends",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_dividend_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<DividendService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_dividend_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}